Add tests

+++ src/MarketData.Gateway/Models/FxQuote.cs
@@ -3,7 +3,7 @@ namespace MarketData.Gateway.Models
     public class FxQuote : MarketData
     {
         public Currency Currency { get; set; }
-        public float Bid { get; set; }
-        public float Ask { get; set; }
+        public decimal Bid { get; set; }
+        public decimal Ask { get; set; }
     }
 }
